backend/requirements.txt
# Core Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# CORS handling
django-cors-headers>=4.3,<5.0

# Environment variables
python-dotenv>=1.0,<2.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
hypothesis>=6.92,<7.0
factory-boy>=3.3,<4.0

# Code quality
flake8>=6.1,<7.0
black>=23.12,<24.0

# Image handling
Pillow>=10.1,<11.0

# Utilities
python-dateutil>=2.8,<3.0
